Fix std::is_convertible for the case where From and To are the same non-copyable type (#89)
Resolves #88 (Fix std::is_convertible for the case where From and To are the same non-copyable type).

std::is_convertible<From, To> previously evaluated to std::false_type if
From and To were the same non-copyable type. Having
std::is_convertible<From, To> evaluate to std::true_type if
std::is_same<From, To> evaluates to std::true_type fixes this, but does
assume that the type is move or copy constructible which may not be the
case.